Prod front end cities admin

+++ backend/src/cities/cities.controller.ts
@@ -1,6 +1,7 @@
 import { Request, Response, NextFunction} from 'express'
 import { orm } from '../shared/db/orm.js'
 import { City } from './cities.entity.js'
+import { error } from 'console'
 
 function sanitizeCityInput(req: Request, res: Response, next: NextFunction) {
   req.body.sanitizedInput = {
@@ -18,11 +19,21 @@ function sanitizeCityInput(req: Request, res: Response, next: NextFunction) {
   next()
 }
 
+function validateCityInput(req: Request, res: Response, next: NextFunction) {
+  if (!req.body.sanitizedInput.nameCity || 
+    !req.body.sanitizedInput.idCity || 
+    !req.body.sanitizedInput.province ||
+    req.body.sanitizedInput.nameCity.trim() === '') {
+    return res.status(400).json({ message: 'Description and Office ID are required.' })
+  }
+  next()
+}
+
 const em = orm.em
 
 async function findAll(req: Request, res: Response) {
   try {
-    const cities = await em.find(City, {})
+    const cities = await em.find(City, {}, {populate: ['province']})
     res.status(200).json({ message: 'find all cities', data: cities })
   }catch (error : any) {
     res.status(500).json({ message: error.message })
@@ -57,7 +68,8 @@ async function update(req: Request, res: Response) {
     const city = await em.findOneOrFail(City,  { idCity : id })
     em.assign(city, req.body.sanitizedInput)
     await em.flush() 
-    res.status(200).json({ message: 'City updated successfully'})
+    const updatedCity = await em.findOneOrFail(City, { idCity: id }, {populate: ['province']})
+    res.status(200).json({ message: 'City updated successfully', data: updatedCity })
   }catch (error: any) {
     res.status(500).json({ message: error.message })
   }
@@ -66,6 +78,9 @@ async function update(req: Request, res: Response) {
 async function remove(req: Request, res: Response) {
   try{
     const id = Number.parseInt(req.params.idCity)
+    if (isNaN(id)) {
+      return res.status(400).json({ message: "Invalid city id" })
+    }
     const city = em.getReference(City, id as any)
     await em.removeAndFlush(city)
     res.status(200).json({ message: 'City removed successfully' })
@@ -74,4 +89,4 @@ async function remove(req: Request, res: Response) {
   }
 }
 
-export {sanitizeCityInput, findAll, findOne, add, update, remove }
+export {sanitizeCityInput, validateCityInput, findAll, findOne, add, update, remove }
